requirements.txt
imageio
matplotlib
scikit-image
easydict
torch==1.1.0
torchvision==0.3.0

path.py
opencv-python>=3.0,<4.0
fast_slic
tensorboardX